Añadiendo servicios y tipos de propiedad a la DDBB
También se modificó la ruta getProperties para que me devuelva los servicios que posee la propiedad
