+++ api/src/controllers/PropertyController.js
@@ -1,5 +1,4 @@
-const { Property } = require("../db/index.js")
-const { Op } = require("sequelize")
+const { Property, Service } = require("../db/index.js")
 
 const getPropertyById = async (req, res, next) => {
   try {
@@ -59,106 +58,19 @@ const addProperty = async (req, res) => {
 
 const getAll = async (req, res, next) => {
   try {
-    let {
-      attributes,
-      order = "ASC",
-      orderBy = "name",
-      search,
-      exclude,
-      minprice = null,
-      maxprice = null,
-      type, // id_type_property
-      location,
-      minrooms = null,
-      maxrooms = null,
-      minpeople = null,
-      maxpeople = null,
-    } = req.query
-    const options = { where: {} }
+      const options = req.options || { where: {}};
 
-    // Definimos que atributos quiere traer o excluir
-    if (attributes || exclude) {
-      options.attributes = attributes
-        ? attributes.split(" ")
-        : { exclude: exclude.split(" ") }
-    }
-
-    // Definiendo el ordenamiendo: default por nombre ascendentemente
-    orderBy = orderBy.split(" ")
-    orderBy.push(order)
-    options.order = [orderBy]
-
-    // Definimos si filtrado por locacion
-    if (location) {
-      options.where = {
-        location: {
-          [Op.iLike]: `%${location}%`,
+    // incluimos los servicios de la propiedad
+    options.include = [
+      {
+        model: Service,
+        attributes: ["name","id"],
+        through: {
+          attributes: [], // que atributos de aquí quiero o si está vacío me elimina el atributo Country anidado
         },
-      }
-    }
-
-    // Definimos si hay un termino de búsqueda
-    if (search) {
-      options.where = {
-        name: {
-          [Op.iLike]: `%${search}%`,
-        },
-      }
-    }
-
-    // Defino si hay filtros por precio
-    if (minprice !== null && maxprice !== null) {
-      options.where.price = {
-        [Op.and]: [{ [Op.gte]: minprice }, { [Op.lte]: maxprice }],
-      }
-    } else if (minprice !== null || maxprice !== null) {
-      options.where.price =
-        minprice !== null
-          ? {
-              [Op.gte]: minprice,
-            }
-          : {
-              [Op.lte]: maxprice,
-            }
-    }
-
-    // Defino si hay filtros por numero de cuartos
-    if (minrooms !== null && maxrooms !== null) {
-      options.where.numberOfRooms = {
-        [Op.and]: [{ [Op.gte]: minrooms }, { [Op.lte]: maxrooms }],
-      }
-    } else if (minrooms !== null || maxrooms !== null) {
-      options.where.numberOfRooms =
-        minrooms !== null
-          ? {
-              [Op.gte]: minrooms,
-            }
-          : {
-              [Op.lte]: maxrooms,
-            }
-    }
-
-    // Defino si hay filtros por numero de personas permitido con rango
-    if (minpeople !== null && maxpeople !== null) {
-      options.where.maxNumberOfPeople = {
-        [Op.and]: [{ [Op.gte]: minpeople }, { [Op.lte]: maxpeople }],
-      }
-    } else if (minpeople !== null || maxpeople !== null) {
-      options.where.maxNumberOfPeople =
-        minpeople !== null
-          ? {
-              [Op.gte]: minpeople,
-            }
-          : {
-              [Op.lte]: maxpeople,
-            }
-    }
-
-    // Defino si hay filtros por typo de propiedad
-    if (type) {
-      options.where.id_type_property = type
-    }
-
+      },
+    ]
+    
     // console.log(options)
     const properties = await Property.findAll(options)
