+++ api/src/data/users.json
@@ -1,5 +1,6 @@
 [
   {
+    "id": "b49a5948-21a0-44c3-92fc-20b626d94dc2",
     "name": "Viviana",
     "lastname": "Calva",
     "country": "Ecuador",
@@ -12,6 +13,7 @@
     "type": "ADMIN"
   },
   {
+    "id": "3ae3b92c-af90-40b4-a5de-8d8f1bb4096e",
     "name": "Maritza",
     "lastname": "López",
     "country": "Peru",
@@ -23,6 +25,7 @@
     "status": "enabled"
   },
   {
+    "id": "b2119946-a958-4744-ad61-f05ce92f7024",
     "name": "Juan",
     "lastname": "Perez",
     "country": "Colombia",
@@ -34,6 +37,7 @@
     "status": "enabled"
   },
   {
+    "id": "79864a10-2446-42e5-ab6d-6867f08eca69",
     "name": "Julieta",
     "lastname": "Gomez",
     "country": "Colombia",
