+++ api/src/index.js
@@ -9,6 +9,6 @@ sequelize.sync({ force: true }).then(() => {
     console.log(`Server running on port ${PORT}`) // eslint-disable-line no-console
 
     // Inicializamos la DDBB
-    // await initDB();
+    await initDB();
   })
 })
